feat(updates)

+++ src/modules/files/file.services.js
@@ -1,32 +1,43 @@
 import { deleteFile } from "../../configs/upload.config.js";
 import FileModel from "./file.models.js";
-
+import { paginate } from "../../utils/paginate.utils.js";
 
 export const handleFileDeletion = async (id) => {
     const file = await FileModel.findById(id);
-    const x = await deleteFile(file.fileUrl);
-    console.log(x)
-}
+    if (!file) throw new Error("File not found");
 
-export const destroy = async (id) => {
-    return await FileModel.findByIdAndDelete(id);
-}
+    await deleteFile(file.fileUrl);
+};
 
-export const update = async (id, data) => {
-    return await FileModel.findByIdAndUpdate(id, data, { returnDocument: 'after' });
-}
+export const destroy = async (id, owner) => {
+    return await FileModel.findOneAndDelete({ _id: id, owner });
+};
 
-export const get = async (id) => {
-    return await FileModel.findById(id);
-}
+export const update = async (id, owner, data) => {
+    return await FileModel.findOneAndUpdate(
+        { _id: id, owner },
+        data,
+        { new: true }
+    );
+};
 
-export const list = async () => {
-    return await FileModel.find().sort("-createdAt");
-}
+export const get = async (id, owner) => {
+    return await FileModel.findOne({ _id: id, owner });
+};
+
+export const list = async (page, limit, filters = {}, search = null) => {
+    return await paginate(FileModel, filters, {
+        page,
+        limit,
+        search,
+        searchFields: [],
+        populate: [
+            { path: "owner", select: "username" },
+            { path: "folder", select: "name" },
+        ]
+    });
+};
 
 export const add = async (data) => {
-    const file = new FileModel(data)
-
-    return await file.save();
-}
-
+    return await FileModel.create(data);
+};
